feat: add slides component
